DEVPROD-9875: Create generator task for parallel e2e tasks
