+++ .evergreen/generate.yml
@@ -6,6 +6,28 @@ functions:
       script: |
         node .evergreen/scripts/generate-tasks.js
 
+  run-generate-parallel-e2e-tasks:
+    command: shell.exec
+    params:
+      working_dir: ui
+      env:
+        BUILD_VARIANT: ${build_variant}
+      script: |
+        node .evergreen/scripts/generate-parallel-e2e-tasks.js
+
+  attach-e2e-generated:
+    command: s3.put
+    type: system
+    params:
+      aws_key: ${AWS_ACCESS_KEY_ID}
+      aws_secret: ${AWS_SECRET_ACCESS_KEY}
+      aws_session_token: ${AWS_SESSION_TOKEN}
+      local_files_include_filter: ["ui/.evergreen/generate-parallel-e2e-tasks.json"]
+      remote_file: ${build_variant}/${task_id}/${execution}/generate-parallel-e2e-tasks.json
+      bucket: evg-bucket-evergreen-ui
+      content_type: text/plain
+      permissions: public-read
+
   attach-generated:
     command: s3.put
     type: system
@@ -28,3 +50,11 @@ tasks:
         params:
           files:
             - ui/.evergreen/generate-tasks.json
+  - name: generate-parallel-e2e-tasks
+    commands:
+      - func: run-generate-parallel-e2e-tasks
+      - func: attach-e2e-generated
+      - command: generate.tasks
+        params:
+          files:
+            - ui/.evergreen/generate-parallel-e2e-tasks.json
